requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dialogflow==0.6.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
google-api-core==1.14.0
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
mongoengine==0.18.2
pkg-resources==0.0.0
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pymongo==3.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
urllib3==1.25.3
Werkzeug==0.15.4
